feat: add ability to display past food diaries
Created "pathvariabled" view to history of food diaries, with different labels when user is on today's view and when on past ones. Views to edit, update and delete meals and products are corrected for new circumstances, so user can update forementioned entries for any given date of diary.

Also set global date format in application.properties to avoid using @DateTimeFormat(pattern = "yyyy-MM-dd") annotation in many places.